refactor: reorganize billing package structure
Moves billing-related classes to a dedicated package for clearer separation of concerns. Updates import paths to match new locations, ensuring consistent package structure and improving code organization.
