+++ src/main/java/sae/semestre/six/bill/BillDetail.java
@@ -1,4 +1,4 @@
-package sae.semestre.six.model;
+package sae.semestre.six.bill;
 
 import jakarta.persistence.*;
