+++ src/test/java/sae/semestre/six/controller/BillingControllerTest.java
@@ -1,6 +1,8 @@
 package sae.semestre.six.controller;
 
 import org.junit.Test;
+import sae.semestre.six.bill.BillingController;
+
 import static org.junit.Assert.*;
 import java.io.File;
